feat : 상수 내용 분리
