+++ caffhheiene/src/constants/previousProjects.ts
@@ -1,0 +1,35 @@
+import { BADGES } from '@/constants/badge';
+
+export const PREVIOUS_PROJECTS = [
+  {
+    link: 'https://github.com/SWARVY/KUBS',
+    name: 'KUBS (경기대학교 종합 예약 시스템)',
+    description: '학교 시설물들에 대한 대여 시스템을 웹에서 제공하는 서비스를 개발하였습니다',
+    stack: [BADGES.react, BADGES.javascript, BADGES.redux],
+  },
+  {
+    link: 'https://github.com/SWARVY/Ezipnaezip_web',
+    name: '이집내집',
+    description: '웹에서 DALL.E를 사용한 AI 인테리어 디자인 / 커뮤니티 서비스를 개발하였습니다',
+    stack: [BADGES.react, BADGES.javascript, BADGES.tailwindcss, BADGES.recoil, BADGES.reactQuery],
+  },
+  {
+    link: 'https://github.com/TeamCommerce',
+    name: '팀커머스',
+    description: '이 커머스 웹 플랫폼(피팅노트)을 클론 코딩하였습니다',
+    stack: [BADGES.react, BADGES.typescript, BADGES.recoil, BADGES.reactQuery, BADGES.tailwindcss],
+  },
+  {
+    link: 'https://github.com/SWARVY/Caffhheiene_web',
+    name: 'Caffhheiene (개인 기술블로그)',
+    description: '지금 보고계시는 이 블로그를 직접 만들었습니다',
+    stack: [
+      BADGES.nextjs,
+      BADGES.typescript,
+      BADGES.recoil,
+      BADGES.reactQuery,
+      BADGES.tailwindcss,
+      BADGES.framer,
+    ],
+  },
+];
